CustomArrowHeadSave

+++ PresentationLayer/Views/UserControls/ICustomArrowHeadUC.cs
@@ -5,7 +5,7 @@ namespace PresentationLayer.Views.UserControls
     public interface ICustomArrowHeadUC
     {
         int ArrowWidthId { get; set; }
-        int Arrow_Size { get; set; }
+        decimal Arrow_Size { get; set; }
         int ArrowCountWD { get; set; }
         int ArrowCountHT { get; set; }
